Renamed the "Moosic" folder to "Music"
