+++ src/main/java/com/godson/kekbot/Music/AudioPlayerSendHandler.java
@@ -1,4 +1,4 @@
-package com.godson.kekbot.Moosic;
+package com.godson.kekbot.Music;
 
 import com.sedmelluq.discord.lavaplayer.player.AudioPlayer;
 import com.sedmelluq.discord.lavaplayer.track.playback.AudioFrame;
