+++ src/main/java/com/godson/kekbot/Music/GuildMusicManager.java
@@ -1,9 +1,8 @@
-package com.godson.kekbot.Moosic;
+package com.godson.kekbot.Music;
 
 import com.darichey.discord.api.CommandContext;
 import com.sedmelluq.discord.lavaplayer.player.AudioPlayer;
 import com.sedmelluq.discord.lavaplayer.player.AudioPlayerManager;
-import net.dv8tion.jda.core.entities.Guild;
 import net.dv8tion.jda.core.entities.TextChannel;
 import net.dv8tion.jda.core.entities.User;
